doc/next: directory for release note fragments
This is the first CL in a sequence that adds support
for generating release notes from fragments.

The actual generator will live elsewhere, in x/build.
This repo will hold the content and some validity
checks.

doc/next will hold the fragments. This CL explains their structure in
a README.

For #64169.

For golang/go#64169.

Change-Id: Iaa8d9ad96393ab9433170b3cfa47334837f3f691