cmd/relnote,doc/next: add release note check
Add a test that every file in api/next has a corresponding
release note fragment.

For now, check that the fragment file has some minimal content.
There is a more thorough check in the x/build/relnote package,
which we'll use when that package is vendored in.

Also, add release note files for the existing API changes
so the test passes.

For #64169.

Change-Id: I40375ab9a8ac20b7e8babd0e69862436ac1238bf