runtime: use span.elemsize for accounting in mallocgc
Currently the final size computed for an object in mallocgc excludes the
allocation header. This is correct in the ASAN case, definitely wrong
for memory profiling, and a little inaccurate for a few of the other
cases (e.g. inittrace). gcmarknewobject also explicitly uses
span.elemsize, because the GC cares about the full size of the object
slot, not just the headerless size.

Make the final size include the malloc header. Clearly more cases care
about span.elemsize than the headerless size. This change adds some
fixup code for only the ASAN case.

For #64153.

Change-Id: I481b34b2bb9ff923b59e8408ab2b8fb9025ba944