cmd/compile/internal/noder: fix compatibility issue with fn.Inl.Properties
When GOEXPERIMENT=newinliner is in effect, an additional "properties"
string is written out as part of the export data for each inlinable
function. To avoid problems with the bootstrapping that happens in
cmd/dist (see the cited bug for more details), switch to a scheme in
which we always write out the properties string, but just leave it
empty in the case that the newinliner experiment is turned off.

Fixes #64189.

Change-Id: Id838c3b316d5ad9c585341dc0a16a06c44e469d2