runtime: add allocalign16 GODEBUG
DO NOT SUBMIT. Needs to be benchmarked. If this causes performance
regressions, just rely on GOEXPERIMENT for switching the alignment back.

This change adds a GODEBUG for backwards compatibility. It ensures that
allocated heap memory that was aligned to 16 bytes before the Go 1.22
release is still aligned, at the cost of some memory use.

Change-Id: I1253d064b46fdbc393755f09f1cf83739b7c4227