net/http: prevent redirect loop in serveFile if "/" is a normal file
When FileServer(Dir("file")) is used where "file" is a normal file and
not a directory, the server enters a redirect loop.

The usage of a file inplace of a directory path is not documented in
http.Dir and it could be considered undefined behavior.

This CL updates (Dir).Open to validate that dir is a directory as well
as updating the documentation accordingly.

One of the tests has a testcase for Dir("/etc/hosts") and it states this
is a trick that ServeFile uses, but I'm not sure if this is still relevant
today.

Fixes #63769

Change-Id: I81e289444e7d0bd72189c2e7b763f5540333e2d0
GitHub-Last-Rev: 6f5f71225053582edc599a866b73962d954a6578
GitHub-Pull-Request: golang/go#63860