cmd/compile/internal/escape: generate tests to execute escape analysis examples in GOROOT/test
WIP. Feedback welcome.

GOROOT/test contains a treasure trove of escape analysis examples and
corner cases.

These are validated today via ERROR comments with expected behavior that
have passed human review, but these can miss a problem (e.g., if a
change author or reviewer misses that an additional log message should
be present, or there is a bug in the logging logic within escape
analysis, and so on).

This CL adds the ability to build and execute these examples, as well as
code generated _test.go files that call almost all of the current
examples in the test/escape*.go files. The runtime today can throw in
some cases if it can determine it found a bad pointer. Later CLs in this
stack attempt to add some additional runtime checks targeting some
violations of two key escape analysis invariants: live stack variables
cannot point to stacks that have returned, and heap cannot point to
stack.

This can serve as an additional backstop for the ERROR comments, which
would presumably continue to be the primary validation mechanism. It
seems worthwhile to do more to test escape analysis, including it might
help give more confidence when making changes.

The testdir harness is also updated to support a new 'test' action,
which is used by these tests. (Compare with the existing 'run' action
implementation). We also ask the types2 tests not to type check these
new _test.go files individually (which would fail).

The intended philosophy is to hopefully avoid being an on-going
maintenance burden, but to cast a wide net while keeping the approach
simple enough that it does not aim for complete coverage (e.g,. some
tests are marked auto-recover for panics).

The code generator has an -autopass mode, which generates the _test.go
files and then immediately runs them to observe any recoverable panics,
timeouts, or other failures, and then generates a final version of the
_test.go files with failing tests marked as skip or auto-recover.

To get a flavor of the results, here are four sample snippets from the
generated reflect_escape_test.go:

  {name: "kind", f: func() { i := gi; kind(i) }},
  {name: "ptr A", f: func() { i := gi; pi := &i; ptr(pi) }},
  {name: "ptr B", f: func() { i := gi; pi := &i; esc(pi); ptr(pi) }},
  {name: "recv1", skip: true, f: func() { c := make(chan string, 10); recv1(c) }},

For the kind test, a local integer i is declared and used to call the
kind func from reflect_escape.go. For the ptr test, an *int is declared,
and there are two generated variations -- one which forces the pointer to
escape, and one which does not. For the recv1 test, the original escape
analysis test in reflect_escape.go hangs forever, so it was
automatically marked skip.

TODO: probably commit the code generator as well, though not sure where
that should live - perhaps src/cmd/internal/genesctests or similar?
TODO: could update escape2.go, escape2n.go, and escape_calls.go to be
buildable, but maybe a later CL. (Currently, they compile, but won't
link due to some missing func bodies).

Change-Id: Idd00f1ff213a12baccd818871dd60512cc4a0162