runtime: refactor fpunwindExpand to use provided buffer
fpunwindExpand currently allocates a new slice to hold the expanded call
stack. In each place it's used, the resulting slice won't be needed
immediately afterward, so the allocation is wasteful. Refactor
fpunwindExpand to instead expand the call stack into a provided buffer.
To avoid making the same change in two places, this CL moves the frame
pointer unwinding-related code into a single place alongside the rest of
traceback-related code.

Change-Id: I05b26c191a8f76404c21ccbe3bd422325540425b