runtime: check for illegal stack pointers beyond sp during GC
WIP. Feedback welcome.

A bug in escape analysis could in theory leave a stack pointer that
points to a frame that has returned. Someone incorrectly using unsafe or
cgo could also end up with a similarly illegal pointer.

Check in scanblock to catch some instances of these.

This check triggers if you purposefully use unsafe to illegally use a
uintptr to return a stack pointer (purposefully blinding escape
analysis):

  runtime: stack pointer at 0xc00004df30 points to stack 0xc00004daf8 beyond sp 0xc00004de88
  runtime: stack [0xc00004c000,0xc00004e000)
  stack obj=0xc00004df20 s.base()=0xc000048000 s.limit=0xc000050000 s.spanclass=0 s.elemsize=8192 s.state=mSpanManual
   *(stack obj+0) = 0x42
   *(stack obj+8) = 0x0
   *(stack obj+16) = 0xc00004daf8 <==
   *(stack obj+24) = 0xc00004daf8
   [...]

Note that this can fairly easily miss an invalid pointer to a small-ish
frame that has returned because the goroutine can be in a runtime stack
that has moved sp sufficiently down again. This check still seems
useful, though. (For example, we have some tests later in this CL stack
that purposefully push the interesting test functions further down the
stack to help this trigger when testing escape analysis logic).

TODO: consider if this should be for debug.invalidptr >= 2 or similar?
TODO: it would be nice to commit a test that triggers this.
Change-Id: I8de75336b321143a1dcd46f1fd8b2329b13d2f7b