net/textproto: improve error handling in skipSpace
The existing implementation assumes that all bufio readers will hold
onto the last read error, however when we introduce a reader with
auto rewind on io.EOF, the error is lost resulting in an infinite loop
in ReadMIMEHeader.  This fix takes a similar approach to bufio and holds
the last read error and returns it if present when calling readLineSlice.

Fixes #53858

Change-Id: Ib13497e7e32be23401222f522dc657427182a7b3