cmd/compile/internal/escape: report why a root outlives a location
Today, the -m=2 output shows a path from an escaping location to a root
location.  In many cases, it is clear why that root escapes, such as
when the root is the heap or prior messages show the root escaping, but
in other cases, it might not be obvious why that root triggers an
escape.

This CL updates the -m=2 output to include a final line that summarizes
why the root shown in the message outlives the escaping location.

This helps differentiate between reasons such as the root outlives a
loop body, the root is a result parameter, or the root outlives a
closure call frame. This message could be helpful in general, and would
have simplified diagnosing CL 546023 in particular.

The final line in this -m=2 output is an example of the new text:

 ks escapes to heap:
   flow: clientKeyShare = &ks:
     from &ks (address-of) at ./handshake_server_tls13.go:189:22
     from clientKeyShare = &ks (assign) at ./handshake_server_tls13.go:189:20
   clientKeyShare declared outside loop and might outlive ks

While here, we also remove (*batch).leakTo, which is unused.

Change-Id: Ibbee4659f6e58aec73478ab32e9f536edf6a1193