cmd/compile/internal/devirtualize: help escape analysis by avoiding function-scoped temporaries
WIP. Feedback welcome.

Today for PGO-based devirtualization, the temporaries that hold the
results of a devirtualized call are declared at function-level scope via
the typecheck.TempAt function.

When the devirtualized call is within a for loop, this means that escape
analysis can unnecessarily mark one or more of the call results as
escaping when it observes a flow from a location within the loop body to
the temporary variable's location declared outside that loop, concluding
that the temporary can outlive the loop.

This CL updates the PGO-based devirtualization to instead declare these
temporaries just before the devirtualized call, which helps escape
analysis more correctly understand the lifetime of these variables.

A follow-on CL tests this via a modification to reflect.Type.Field that
can now eliminate some allocations after this change.

TODO: would be good to have a specific test in this CL as well.
TODO: double-check for possible similar issue with inputs, etc.

Change-Id: I3cb79524aa493d72a509159956e7072870e99876