cmd/link: use types (and not GC programs) to build data/bss ptrmasks
The linker knows the types of the global variables. We can use those
types to build the GC programs that describe the data and bss pointer masks.

That way we don't use the GC programs of the constituent types.
This is part of the effort to remove GC programs altogether.

There's a major complication in that when we're linking against a
shared library (typically, libstd.so), the relocations we need to
break apart arrays and struct to constituent types are difficult to
find. Load that additional data when linking against shared libraries.

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: I8516b24a0604479895c7b8a8a358d3cd8d421530