testing: add regression tests for reentrant calls to T.Run
These tests represent two patterns of usage, found in Google-internal
tests, that deadlocked after CL 506755.

TestConcurrentRun is a minor variation on TestParallelSub, with the additional
expectation that the concurrent calls to Run
(without explicit calls to Parallel) proceed without blocking.
It replaces TestParallelSub.

TestParentRun is similar, but instead of calling Run concurrently it calls Run
from within the subtest body. It almost certainly represents an accidental misuse
of T.Run, but since that pattern used to run to completion we don't want to break
it accidentally. (Perhaps it should be diagnosed with a vet check instead?)

Fixes #64402.

Change-Id: I14fc7e7085a994c284509eac28190c3a8feb04cd